requirements.txt
Django==3.1
django-ckeditor==6.2.0
django-cors-headers==3.6.0
django-js-asset==2.0.0
django-jet
xlwt
gunicorn
gevent
requests
django-adminlte-ui
pymysql
django-treebeard==4.5.1
django-adminlte-ui==1.6.0
psycopg2-binary